Fix user ava

See merge request sli-usth/sli-home-backend!14

+++ src/modules/user/user.service.ts
@@ -54,24 +54,34 @@ export class UserService {
     if (!user) {
       throw new NotFoundException('User does not exist');
     }
-
+  
     if (id !== currentUser.id) {
       throw new ForbiddenException('You do not have permission');
     }
-
-    if (user.avatar) {
-      fs.unlinkSync(user.avatar);
-
-      user.avatar = null;
+  
+    try {
+      // Check and delete the existing avatar if it exists
+      if (user.avatar && fs.existsSync(user.avatar)) {
+        fs.unlinkSync(user.avatar);
+      }
+    } catch (error) {
+      console.error('Error deleting avatar:', error);
+      // Handle the error or log it appropriately
     }
-
+  
+    // Update user information
     user.name = updateUserDto.name;
     user.email = updateUserDto.email;
     user.phoneNumber = updateUserDto.phoneNumber;
-    user.avatar = avatar;
+  
+    // Update avatar only if provided
+    if (avatar) {
+      user.avatar = avatar;
+    }
+  
     return this.userRepo.save(user);
   }
-
+  
   async updatePassword(
     id: number,
     updatePasswordDto: UpdatePasswordDto,
